chore: Poll for data to sync instead of sleeping
