+++ suite/src/__tests__/fast/composedb-model-sync.test.ts
@@ -1,12 +1,11 @@
 import { ComposeClient } from '@composedb/client'
 import { beforeAll, describe, test, expect } from '@jest/globals'
 import { Composite } from '@composedb/devtools'
-import { newCeramic } from '../../utils/ceramicHelpers.js'
+import { loadDocumentOrTimeout, newCeramic } from '../../utils/ceramicHelpers.js'
 import { createDid } from '../../utils/didHelper.js'
 import { BasicSchema } from '../../graphql-schemas/basicSchema'
 import { StreamID } from '@ceramicnetwork/streamid'
 import { waitForDocument } from '../../utils/composeDbHelpers.js'
-import { CommonTestUtils as TestUtils } from '@ceramicnetwork/common-test-utils'
 
 const ComposeDbUrls = String(process.env.COMPOSEDB_URLS).split(',')
 const adminSeeds = String(process.env.COMPOSEDB_ADMIN_DID_SEEDS).split(',')
@@ -40,14 +39,10 @@ describe('Sync Model and ModelInstanceDocument using ComposeDB GraphQL API', () 
     const parts = String(resources[0]).split('model=')
     const modelId = parts[parts.length - 1]
 
-    await TestUtils.waitForConditionOrTimeout(async () =>
-      ceramicInstance2
-        .loadStream(modelId)
-        .then((_) => true)
-        .catch((_) => false),
-    )
+    // Wait for model to be available on node 2.
+    await loadDocumentOrTimeout(ceramicInstance2, StreamID.fromString(modelId), 30 * 1000)
 
-    // start indexing for tha nodel on node 2
+    // start indexing for the model on node 2
     await ceramicInstance2.admin.startIndexingModels([StreamID.fromString(modelId)])
     composeClient2 = await new ComposeClient({
       ceramic: ceramicInstance2,
